+++ Tests/PUWebPaymentsTests/Sources/Models/WebPaymentsResultTests.swift
@@ -21,6 +21,9 @@ final class WebPaymentsResultTests: XCTestCase {
 
     XCTAssertEqual(WebPaymentsResult(status: .externalApplication, url: url).status, .externalApplication)
     XCTAssertEqual(WebPaymentsResult(status: .externalApplication, url: url).url, url)
+    
+    XCTAssertEqual(WebPaymentsResult(status: .externalBrowser, url: url).status, .externalBrowser)
+    XCTAssertEqual(WebPaymentsResult(status: .externalBrowser, url: url).url, url)
   }
 
 }