+++ Tests/PUWebPaymentsTests/Sources/WebPaymentsViewModelTests.swift
@@ -51,14 +51,14 @@ final class WebPaymentsViewModelTests: XCTestCase {
   func testNavigationPolicyForUrlShouldInformDelegateThatCurrentUrlDidUpdate() throws {
     let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
     given(matcher.result(any())).willReturn(.notMatched)
-    _ = sut.navigationPolicy(url)
+    _ = sut.navigationPolicy(for: url, inMainFrame: true)
     verify(delegate.webPaymentsViewModel(any(), didUpdate: url)).wasCalled()
   }
 
   func testNavigationPolicyForUrlReturnsAllowForNotMatchedMatcherResult() throws {
     let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
     given(matcher.result(any())).willReturn(.notMatched)
-    XCTAssertEqual(sut.navigationPolicy(url), .allow)
+    XCTAssertEqual(sut.navigationPolicy(for: url, inMainFrame: true), .allow)
     verify(matcher.result(url)).wasCalled()
     verify(
       delegate
@@ -71,7 +71,7 @@ final class WebPaymentsViewModelTests: XCTestCase {
   func testNavigationPolicyForUrlReturnsCancelForSuccessMatcherResult() throws {
     let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
     given(matcher.result(any())).willReturn(.success)
-    XCTAssertEqual(sut.navigationPolicy(url), .cancel)
+    XCTAssertEqual(sut.navigationPolicy(for: url, inMainFrame: true), .cancel)
     verify(matcher.result(url)).wasCalled()
     verify(
       delegate
@@ -86,7 +86,7 @@ final class WebPaymentsViewModelTests: XCTestCase {
   func testNavigationPolicyForUrlReturnsCancelForFailureMatcherResult() throws {
     let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
     given(matcher.result(any())).willReturn(.failure)
-    XCTAssertEqual(sut.navigationPolicy(url), .cancel)
+    XCTAssertEqual(sut.navigationPolicy(for: url, inMainFrame: true), .cancel)
     verify(matcher.result(url)).wasCalled()
     verify(
       delegate
@@ -101,7 +101,7 @@ final class WebPaymentsViewModelTests: XCTestCase {
   func testNavigationPolicyForUrlReturnsAllowForContinue3DSMatcherResult() throws {
     let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
     given(matcher.result(any())).willReturn(.continue3DS)
-    XCTAssertEqual(sut.navigationPolicy(url), .allow)
+    XCTAssertEqual(sut.navigationPolicy(for: url, inMainFrame: true), .allow)
     verify(matcher.result(url)).wasCalled()
     verify(
       delegate
@@ -116,7 +116,7 @@ final class WebPaymentsViewModelTests: XCTestCase {
   func testNavigationPolicyForUrlReturnsAllowForContinueCVVMatcherResult() throws {
     let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
     given(matcher.result(any())).willReturn(.continueCvv)
-    XCTAssertEqual(sut.navigationPolicy(url), .allow)
+    XCTAssertEqual(sut.navigationPolicy(for: url, inMainFrame: true), .allow)
     verify(matcher.result(url)).wasCalled()
     verify(
       delegate
@@ -131,7 +131,7 @@ final class WebPaymentsViewModelTests: XCTestCase {
   func testNavigationPolicyForUrlReturnsCancelForExternalApplicationMatcherResult() throws {
     let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
     given(matcher.result(any())).willReturn(.externalApplication)
-    XCTAssertEqual(sut.navigationPolicy(url), .cancel)
+    XCTAssertEqual(sut.navigationPolicy(for: url, inMainFrame: true), .cancel)
     verify(matcher.result(url)).wasCalled()
     verify(
       delegate
@@ -141,6 +141,32 @@ final class WebPaymentsViewModelTests: XCTestCase {
             status: .externalApplication,
             url: sut.currentUrl)))
     .wasCalled()
+  }
+    
+  func testNavigationPolicyForUrlReturnsCancelAndPresentsDialogForCreditExternalApplicationMatcherResult() throws {
+    let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
+    given(matcher.result(any())).willReturn(.creditExternalApplication)
+    XCTAssertEqual(sut.navigationPolicy(for: url, inMainFrame: true), .cancel)
+    verify(matcher.result(url)).wasCalled()
+    verify(
+      delegate
+        .webPaymentsViewModelShouldPresentProviderRedirectDialog(
+          any(),
+          url))
+    .wasCalled()
+  }
+  
+  func testNavigationPolicyForNotMatchedUrlOutsideMainFrameDoesntUpdateAddressBar() throws {
+    let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
+    given(matcher.result(any())).willReturn(.notMatched)
+    XCTAssertEqual(sut.navigationPolicy(for: url, inMainFrame: false), .allow)
+    verify(matcher.result(url)).wasCalled()
+    verify(
+      delegate
+        .webPaymentsViewModelShouldPresentProviderRedirectDialog(
+          any(),
+          url))
+    .wasNeverCalled()
   }
 
   func testDidTapBackShouldCallDelegateToPresentBackAlertDialog() throws {
@@ -172,5 +198,30 @@ final class WebPaymentsViewModelTests: XCTestCase {
     sut.didFailNavigation(with: ErrorMock.error)
     verify(delegate.webPaymentsViewModel(any(), didUpdate: sut.currentUrl)).wasCalled()
     verify(delegate.webPaymentsViewModelShouldPresentSSLAlertDialog(any())).wasCalled()
+  }
+    
+  func testDidProceedWithInstallmentsExternalApplicationShouldCompleteWithExternalApplicationStatus() throws {
+    let url = URL(string: "https://www.payu.com?authenticationId=\(UUID().uuidString)")!
+    sut.didProceedWithCreditExternalApplication(url)
+      verify(
+        delegate
+          .webPaymentsViewModel(
+            any(),
+            didComplete: WebPaymentsResult(
+              status: .externalBrowser,
+              url: sut.currentUrl)))
+      .wasCalled()
+  }
+    
+  func testDidAbortInstallmentsExternalApplicationShouldCompleteWithExternalApplicationStatus() throws {
+    sut.didAbortCreditExternalApplication()
+      verify(
+        delegate
+          .webPaymentsViewModel(
+            any(),
+            didComplete: WebPaymentsResult(
+              status: .cancelled,
+              url: sut.currentUrl)))
+      .wasCalled()
   }
 }